requirements.txt
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
numpy
matplotlib
tqdm
scikit-learn
pillow